more architecture changes (moved ui parts) but navigation not working
