+++ functionality/shared/met-api/build.gradle
@@ -47,6 +47,8 @@ dependencies {
     // Hilt
     implementation "com.google.dagger:hilt-android:2.44"
     kapt "com.google.dagger:hilt-compiler:2.44"
+
+
 }
 
 // Allow references to generated code
