+++ functionality/shared/met-api/src/main/java/com/example/met_api/BaseItem.kt
@@ -1,0 +1,3 @@
+package com.example.met_api
+
+abstract class BaseItem()
