+++ ui/common/build.gradle
@@ -39,6 +39,8 @@ dependencies {
     implementation 'androidx.core:core-ktx:1.8.0'
     implementation 'androidx.appcompat:appcompat:1.6.1'
     implementation 'com.google.android.material:material:1.9.0'
+    implementation 'androidx.navigation:navigation-fragment-ktx:2.5.3'
+    implementation 'androidx.navigation:navigation-ui-ktx:2.5.3'
     testImplementation 'junit:junit:4.13.2'
     androidTestImplementation 'androidx.test.ext:junit:1.1.5'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.1'
@@ -46,6 +48,5 @@ dependencies {
     // Coil
     implementation "io.coil-kt:coil:1.1.1"
 
-    api project (':ui:detail')
-    api project (':ui:search')
+    api project (":functionality:shared:met-api")
 }
