+++ ui/common/src/main/res/drawable/baseline_broken_image_24.xml
