+++ ui/common/src/main/res/drawable/loading_animation.xml
