+++ ui/common/src/main/res/drawable/loading_image.xml
