+++ ui/common/src/main/res/values/strings.xml
@@ -1,0 +1,8 @@
+<resources>
+    <string name="app_name">MET Gallery</string>
+    <string name="search_hint">Search for Art</string>
+    <string name="artist">Created by %1$s</string>
+    <string name="culture">Culture: %1$s</string>
+    <string name="department">Department: %1$s</string>
+    <string name="more_pictures">More pictures:</string>
+</resources>
