+++ ui/detail/build.gradle
@@ -51,7 +51,8 @@ dependencies {
     implementation "androidx.navigation:navigation-fragment-ktx:2.5.2"
     implementation "androidx.navigation:navigation-ui-ktx:2.5.2"
 
-    api project (':app')
+    api project (":ui:common")
     api project (":functionality:shared:met-api")
     api project (':functionality:detail:entities')
+    api project (":functionality:detail")
 }
