+++ ui/detail/src/main/java/com/example/detail/state/DetailViewModel.kt
@@ -4,7 +4,7 @@ import androidx.lifecycle.SavedStateHandle
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.example.detail.GetDetailsUseCase
-import com.example.metmuseum.Result
+import com.example.met_api.Result
 import com.example.detail.toUiModel
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.SharingStarted
