+++ ui/detail/src/main/res/layout/grid_view_item.xml
