+++ ui/search/build.gradle
@@ -52,6 +52,7 @@ dependencies {
     implementation "androidx.navigation:navigation-fragment-ktx:2.5.2"
     implementation "androidx.navigation:navigation-ui-ktx:2.5.2"
 
-    api project (':app')
     api project (":functionality:shared:met-api")
+    api project (":functionality:search")
+    api project (":ui:common")
 }
