+++ ui/search/src/main/res/layout/fragment_gallery.xml
