+++ ui/search/src/main/res/layout/list_view_item.xml
